fix: comprehensive end-to-end audit — 28 files, 18 verified fixes
Backend:
- config.py: stable dev SECRET_KEY (no per-process random), cloud env warning
- database.py: PostgreSQL pool config (pool_size, max_overflow, pool_pre_ping),
  check_db_connection() for real health checks
- main.py: health endpoint checks DB (returns 503 if down), SlowAPIMiddleware
  for proper rate limiting across all routes
- auth_routes.py: removed separate Limiter instance (was causing 500 on 429),
  rate limiting now handled by shared app.state.limiter via SlowAPIMiddleware
- matching_engine.py: fix N+1 queries — JOIN instead of per-volunteer DB query
- volunteer_routes.py: fix N+1 in list_volunteers — single JOIN query
- user_routes.py: prevent admin self-demotion (would leave zero admins),
  add response_model to change_password endpoint
- schemas/user.py: AdminCreateUser password policy matches UserCreate
  (uppercase + digit required), remove dead TokenData class
- schemas/volunteer.py: JSON validation for availability field,
  rating changed to float for accurate running average
- schemas/community_need.py: add validators to CommunityNeedUpdate
  (negative affected_people, lat/lng range)
- models/volunteer.py: rating column Float (was Integer, lost precision)
- dashboard_routes.py: fix double imports in get_impact_stats,
  add hours_spent.isnot(None) filter, fix resolved_needs to use RESOLVED status
- needs_routes.py: remove wasted count() query on every list request
- seed_prod.py: warn on all cloud providers (not just Render),
  validate password meets policy before creating admin
- requirements.txt: remove unused httpx, upgrade bcrypt to 4.1.3

Frontend:
- api/client.ts: add 15s timeout, fix 401 redirect loop on /login page
- AdminPage.tsx: remove dead activeTab state, fix password minLength 6->8,
  add role change modal, add password reset modal, add deactivation confirmation
- DashboardPage.tsx: fix AnimatedNumber (re-animates on value change),
  add .catch() to Promise.all, show real last-updated time, add error state
- ImpactPage.tsx: fix wrong 'Needs Resolved' metric (was total-open, now
  uses /dashboard/impact endpoint with proper resolved_needs count),
  remove unused Clock import, add monthly resolved area chart
- MapPage.tsx: increase limit 100->500, fix FitBounds dep (coords key not
  length), improve no-coordinates message
- NeedsPage.tsx: fix double fetchNeeds on mount, fix stale closure in
  second useEffect, add lat/lng fields to create form, fix state->undefined
- TasksPage.tsx: separate needs fetch (once on mount, not every page change),
  fix double fetchTasks on mount
- ProfilePage.tsx: fix stale closure (useEffect deps include user.role/id),
  fix token null assertion, refresh profile after task completion
- InstallPrompt.tsx: add try/catch around prompt(), match app dark theme
- UpdatePrompt.tsx: remove console.log in production, match app dark theme
- vite.config.ts: remove missing screenshot assets from PWA manifest,
  remove missing masked-icon.svg from includeAssets
- package.json: move @types/leaflet to devDependencies, bump version to 2.0.0
- NeedsMap.tsx: deleted (dead code, never imported, had latitude:0 bug)

+++ backend/src/models/volunteer.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, String, Boolean, DateTime, Integer, Text, ForeignKey
+from sqlalchemy import Column, String, Boolean, DateTime, Integer, Float, Text, ForeignKey
 from sqlalchemy.orm import relationship
 from sqlalchemy.sql import func
 import uuid
@@ -18,7 +18,7 @@ class VolunteerProfile(Base):
     is_available = Column(Boolean, default=True)
     total_tasks_completed = Column(Integer, default=0)
     total_hours_contributed = Column(Integer, default=0)  # cumulative volunteer hours
-    rating = Column(Integer, default=0)            # 0-5 average rating
+    rating = Column(Float, default=0.0)            # 0.0–5.0 running average rating
     created_at = Column(DateTime, server_default=func.now(), nullable=False)
     updated_at = Column(DateTime, server_default=func.now(), onupdate=func.now(), nullable=False)
